requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
Authlib==1.3.1
beautifulsoup4==4.12.3
blinker==1.8.2
blis==1.0.1
bs4==0.0.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
comm==0.2.2
confection==0.1.5
cryptography==43.0.3
cymem==2.0.8
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
deepdiff==6.7.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
exceptiongroup==1.2.2
executing==2.1.0
Flask==3.0.3
frozenlist==1.5.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.67.1
grpcio-health-checking==1.67.1
grpcio-status==1.67.0
grpcio-tools==1.67.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.0
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.6
langchain-community==0.3.4
langchain-core==0.3.15
langchain-google-genai==2.0.4
langchain-text-splitters==0.3.1
langcodes==3.4.1
langsmith==0.1.138
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mesop==0.12.7
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
narwhals==1.9.4
nest-asyncio==1.6.0
nltk==3.8.1
numpy==1.26.4
ordered-set==4.1.0
orjson==3.10.10
packaging==24.1
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.6
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.2
rpds-py==0.20.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.36
srsly==2.4.8
stack-data==0.6.3
streamlit==1.39.0
tenacity==9.0.0
textblob==0.18.0.post0
thinc==8.3.2
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vaderSentiment==3.3.2
validators==0.34.0
wasabi==1.1.3
watchdog==5.0.3
wcwidth==0.2.13
weasel==0.4.1
weaviate-client==4.9.0
Werkzeug==3.0.6
wrapt==1.16.0
yarl==1.17.1
zipp==3.20.2
